+++ apikeys/index.js
@@ -12,6 +12,8 @@ var JWS = rs.jws.JWS;
 var requestLib = require("request");
 var _ = require("lodash");
 
+const AuthorizationHelper = require('../lib/AuthorizationHelper');
+
 const PRIVATE_JWT_VALUES = ["application_name", "client_id", "api_product_list", "iat", "exp"];
 
 const acceptAlg = ["RS256"];
@@ -24,10 +26,12 @@ var cacheKey = false;
 
 const LOG_TAG_COMP = 'apikeys';
 const CONSOLE_LOG_TAG_COMP = 'microgateway-plugins apikeys';
+let authorizationHelper = null;
 
 module.exports.init = function(config, logger, stats) {
 
     if (config === undefined || !config) return (undefined);
+    authorizationHelper = new AuthorizationHelper(debug);
 
     var request = config.request ? requestLib.defaults(config.request) : requestLib;
     var keys = config.jwk_keys ? JSON.parse(config.jwk_keys) : null;
@@ -178,7 +182,7 @@ module.exports.init = function(config, logger, stats) {
         debug(decodedToken)
         if (keys) {
             debug("using jwk");
-            var pem = getPEM(decodedToken, keys);
+            var pem = authorizationHelper.getPEM(decodedToken, keys);
             try {
                 isValid = JWS.verifyJWT(oauthtoken, pem, acceptField);
             } catch (error) {
@@ -245,49 +249,11 @@ module.exports.init = function(config, logger, stats) {
 
 }
 
-function getPEM(decodedToken, keys) {
-    var i = 0;
-    debug("jwk kid " + decodedToken.headerObj.kid);
-    for (; i < keys.length; i++) {
-        if (keys.kid === decodedToken.headerObj.kid) {
-            break;
-        }
-    }
-    var publickey = rs.KEYUTIL.getKey(keys.keys[i]);
-    return rs.KEYUTIL.getPEM(publickey);
-}
-
-function setResponseCode(res,code) {
-    switch ( code ) {
-        case 'invalid_request': {
-            res.statusCode = 400;
-            break;
-        }
-        case 'access_denied':{
-            res.statusCode = 403;
-            break;
-        }
-        case 'invalid_token':
-        case 'missing_authorization':
-        case 'invalid_authorization': {
-            res.statusCode = 401;
-            break;
-        }
-        case 'gateway_timeout': {
-            res.statusCode = 504;
-            break;
-        }
-        default: {
-            res.statusCode = 500;
-            break;
-        }
-    }
-}
 
 
 function sendError(req, res, next, logger, stats, code, message) {
 
-    setResponseCode(res,code)
+    authorizationHelper.setResponseCode(res,code)
 
     var response = {
         error: code,
